linkedin_scrape/requirements.txt
numpy==1.24.3
pandas==2.0.3
selenium==4.16.0
beautifulsoup4==4.12.0
sqlalchemy==2.0.0
psycopg2-binary==2.9.6
schedule==1.2.0
python-dotenv==1.0.0
requests==2.31.0